+++ Tests/TuistKitTests/Commands/GenerateCommandTests.swift
@@ -2,7 +2,7 @@ import Basic
 import Foundation
 import TuistCore
 import Utility
-import xcodeproj
+import XcodeProj
 import XCTest
 @testable import TuistCoreTesting
 @testable import TuistKit