+++ fixtures/ios_app_with_transitive_framework/Framework1/Project.swift
@@ -10,7 +10,7 @@ let project = Project(
                infoPlist: "Config/Framework1-Info.plist",
                sources: "Sources/**",
                dependencies: [
-                   .framework(path: "../Framework2/Framework2.framework"),
+                   .framework(path: "../Framework2/prebuilt/Framework2.framework"),
         ]),
     ]
 )